refactor: split Program.cs into DI, pipeline, endpoints and hosted service
Program.cs now only bootstraps Serilog and the host. Service registrations
moved to DependencyInjection, middleware/static files/fallback to
PipelineExtensions, and routes split by domain into the Endpoints folder.
The fire-and-forget startup sync Task.Run becomes a proper BackgroundService
(CurrencyStartupSync) registered in DI, with CurrencyApiOptions registered
as a singleton so it can be resolved.
